TGBOT-78 Refactoring: Repetitive code is placed in a separate class
